requirements.txt
fastapi
uvicorn
tweepy
openai
python-dotenv
pandas
matplotlib
seaborn
scikit-learn